+++ packages/mobile-visualization/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## Unreleased
+
+#### 📘 Misc
+
+- Update outdated doc links. [[#440](https://github.com/coinbase/cds/pull/440)]
+
 ## 3.4.0-beta.19 (2/20/2026 PST)
 
 #### 🚀 Updates